+++ aic_scoring/include/aic_scoring/TierScore.hh
@@ -36,7 +36,7 @@ public:
 
   std::string message;
 
-  virtual int total_score() const = 0;
+  virtual double total_score() const = 0;
 
   virtual YAML::Node to_yaml() const {
     YAML::Node score;
@@ -64,7 +64,7 @@ public:
     }
   }
 
-  int total_score() const override {
+  double total_score() const override {
     return score;
   }
 };
@@ -73,18 +73,18 @@ public:
 class Tier2Score : public TierScore {
 public:
   struct CategoryScore {
-    int score;
+    double score;
     std::optional<std::string> message;
 
-    CategoryScore(int s, const std::optional<std::string>& msg) : score(s), message(msg) {}
+    CategoryScore(double s, const std::optional<std::string>& msg) : score(s), message(msg) {}
   };
 
   using CategoryScores = std::map<std::string, CategoryScore>;
 
   Tier2Score(const std::string& msg) : TierScore(msg) {}
 
-  int total_score() const override {
-    int score = 0;
+  double total_score() const override {
+    double score = 0;
     for (const auto& category : category_scores) {
       score += category.second.score;
     }
@@ -104,7 +104,7 @@ public:
     return score;
   }
 
-  void add_category_score(const std::string& category, int score,
+  void add_category_score(const std::string& category, double score,
                           const std::optional<std::string>& msg = std::nullopt) {
     this->category_scores.insert({category, CategoryScore(score, msg)});
   }
@@ -122,12 +122,12 @@ private:
 
 class Tier3Score : public TierScore {
 private:
-  int score;
+  double score;
 
 public:
-  Tier3Score(int s, const std::string& msg) : TierScore(msg), score(s) { }
+  Tier3Score(double s, const std::string& msg) : TierScore(msg), score(s) { }
 
-  int total_score() const override {
+  double total_score() const override {
     return score;
   }
 };
